+++ application/controllers/Assembly.php
@@ -27,46 +27,37 @@ class Assembly extends Application
 
     public function index()
     {
-        $allParts = $this->partsdata->getAllParts();
+        // get user roles
+        $user_role = $this->session->userdata('userrole');
 
-        $this->generateTable($allParts);
-        /*
-          error_reporting(0);
-          // Retrieves the file names for each part in the Robots model.
-          $this->data['head1'] = $this->partsdata->getParts('a', 1)[0]['file_name'];
-          $this->data['body1'] = $this->partsdata->getParts('a', 2)[0]['file_name'];
-          $this->data['feet1'] = $this->partsdata->getParts('a', 3)[0]['file_name'];
-
-          $this->data['head2'] = $this->partsdata->getParts('b', 1)[0]['file_name'];
-          $this->data['body2'] = $this->partsdata->getParts('b', 2)[0]['file_name'];
-          $this->data['feet2'] = $this->partsdata->getParts('b', 3)[0]['file_name'];
-
-          $this->data['head3'] = $this->partsdata->getParts('c', 1)[0]['file_name'];
-          $this->data['body3'] = $this->partsdata->getParts('c', 2)[0]['file_name'];
-          $this->data['feet3'] = $this->partsdata->getParts('c', 3)[0]['file_name'];
-
-          $this->data['head4'] = $this->partsdata->getParts('m', 1)[0]['file_name'];
-          $this->data['body4'] = $this->partsdata->getParts('m', 2)[0]['file_name'];
-          $this->data['feet4'] = $this->partsdata->getParts('m', 3)[0]['file_name'];
-
-          $this->data['head5'] = $this->partsdata->getParts('r', 1)[0]['file_name'];
-          $this->data['body5'] = $this->partsdata->getParts('r', 2)[0]['file_name'];
-          $this->data['feet5'] = $this->partsdata->getParts('r', 3)[0]['file_name'];
-
-          $this->data['head6'] = $this->partsdata->getParts('w', 1)[0]['file_name'];
-          $this->data['body6'] = $this->partsdata->getParts('w', 2)[0]['file_name'];
-          $this->data['feet6'] = $this->partsdata->getParts('w', 3)[0]['file_name'];
-          $this->data['pagebody'] = 'assemblypage';
-          $this->render();
-         * 
-         */
+        // only allow to boss and supervisor
+        if ($user_role == 'boss' || $user_role == 'supervisor')
+        {
+            $allParts = $this->partsdata->getAllSortedParts();
+            $this->generateTable($allParts);
+        } else
+        {
+            $this->data['pagetitle'] = 'Assemble Robot - Boss, Supervisor';
+            $this->data['pagebody'] = 'blockedpage';
+            $this->data['message'] = '';
+            $this->render();
+        }
     }
 
     public function assembleBots()
     {
+        // get parts data
         $head = $this->input->post('head');
         $torso = $this->input->post('torso');
         $legs = $this->input->post('legs');
+
+        // check parts data if one of them is null 
+        if ($head == NULL || $torso == NULL || $legs == NULL)
+        {
+            $this->data['pagebody'] = 'assemblypage';
+            $this->data['message'] = 'Please pick each part once';
+            redirect('/assembly');
+        }
 
         $tempHead = explode('-', $head);
         $tempTorso = explode('-', $torso);
@@ -82,74 +73,49 @@ class Assembly extends Application
 
         $newRobotForInsert = array(
             'head' => $headId,
-            'body' => $torsoId,
+            'torso' => $torsoId,
             'legs' => $legsId
         );
 
         $newRobot = array(
             'head' => $headId,
-            'body' => $torsoId,
+            'torso' => $torsoId,
             'legs' => $legsId,
             'headModel' => $headModel,
             'torsoModel' => $torsoModel,
             'legsModel' => $legsModel,
         );
 
-        // move to model later
-        $this->db->insert('assembledBots', $newRobotForInsert);
-        // delete parts
+        //create a robot
+        $this->robotsdata->createBot($newRobotForInsert);
 
+        // delete the data from parts
+        $this->partsdata->deletePartById($headId);
+        $this->partsdata->deletePartById($torsoId);
+        $this->partsdata->deletePartById($legsId);
+
+        // add to history
         $assembledRobot = $this->createHistory($newRobot, 'Assemble', 0);
-        
+
         // should make a bots history fucntion later
         $this->historydata->insertPartsHistory($assembledRobot);
         redirect('/assembly');
     }
 
-    public function insertAssembled()
-    {
-        if ($this->input->is_ajax_request())
-        { // just additional, to make sure request is from ajax
-            if ($this->input->post('submit_a'))
-            {
-                $head = $this->input->post('head');
-                $body = $this->input->post('body');
-                $legs = $this->input->post('legs');
-                // to model
-                $headId = $this->partsdata->getPartByFile($head)[0]['id'];
-                $bodyId = $this->partsdata->getPartByFile($body)[0]['id'];
-                $legsId = $this->partsdata->getPartByFile($legs)[0]['id'];
-                //
-
-
-                $idNum = $this->robotsdata->getIdNum();
-                $data = array('id' => $idNum, 'head' => $head, 'body' => $body, 'legs' => $legs);
-                $this->robotsdata->createBot($data);
-                $this->partsdata->deletePartById($headId);
-                $this->partsdata->deletePartById($bodyId);
-                $this->partsdata->deletePartById($legsId);
-                $data2 = array('seq'=>$headId.' '.$bodyId.' '.$legsId, 'plant'=>'strawberry', 'model'=>$head.' '.$body.' '.$legs, 'action'=>'Assembly', 'quantity'=>1, 'stamp'=>date('Y-m-d H:i:s'));              
-                $this->robotsdata->updateHistory($data2);
-                // return to view
-                echo json_encode(array("error" => false));
-                return false;
-            }
-        }
-    }
-
+    // create history data
     private function createHistory($part, $action, $amount)
     {
         $temp_array = array();
 
+        // get num of parts
         $num_of_parts = count($part);
 
         $sequence = '';
         $models = '';
-        
-        var_dump($part);
-        $sequence .= $part['head'] . ' ' . $part['body'] . ' ' . $part['legs'];
+
+        // get names
+        $sequence .= $part['head'] . ' ' . $part['torso'] . ' ' . $part['legs'];
         $models .= $part['headModel'] . ' ' . $part['torsoModel'] . ' ' . $part['legsModel'];
-        
 
         $temp_array[] = array(
             'action' => $action,
@@ -167,16 +133,15 @@ class Assembly extends Application
     // generate table of parts
     public function generateTable($allParts)
     {
-        $this->data['pagetitle'] = 'Assembly';
+        $this->data['pagetitle'] = 'Assemble Robot';
         $this->data['pagebody'] = 'assemblypage';
-        $this->data['message'] = "<div></div>";
+        $this->data['message'] = 'Please pick each part once';
 
         $head_parts = array();
         $torso_parts = array();
         $legs_parts = array();
 
         // save parts by piece - head, torso, legs
-
         foreach ($allParts as $part)
         {
             switch ($part['piece'])
@@ -223,6 +188,5 @@ class Assembly extends Application
         $this->data['torso_parts'] = $torso_parts;
         $this->data['legs_parts'] = $legs_parts;
         $this->render();
-    }
-
+    }  
 }